% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{add_metadata}
\alias{add_metadata}
\title{Add Metadata}
\usage{
add_metadata(metadata, title, data)
}
\arguments{
\item{metadata}{The input metadata list}

\item{title}{The title or descriptor of the metadata}

\item{data}{The output file from the Victor run}
}
\value{
An updated list of metadata is returned
}
\description{
Adds additional metadata to the metadata list
}
