% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{ap1_cols}
\alias{ap1_cols}
\title{AP1 Expression Test columns}
\format{An object of class \code{character} of length 2.}
\usage{
ap1_cols
}
\description{
Use these columns when testing AP1 expression.
\describe{
  \item{Control}{control well}
  \item{10 ng/ul PMA}{experiemental well}
}
}
\keyword{datasets}
