% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{cre_cols}
\alias{cre_cols}
\title{CRE Expression Test columns}
\format{An object of class \code{character} of length 2.}
\usage{
cre_cols
}
\description{
Use these columns when testing CRE expression.
\describe{
  \item{Control}{control well}
  \item{10 uM Forskolin}{experiemental well}
}
}
\keyword{datasets}
