% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{expression_barplot}
\alias{expression_barplot}
\title{Expression Barplot}
\usage{
expression_barplot(assay_data)
}
\arguments{
\item{assay_data}{The assay data frame.}
}
\description{
A barplot of expression data.
}
