% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_assay_data}
\alias{get_assay_data}
\title{Get Assay Data}
\usage{
get_assay_data(file)
}
\arguments{
\item{file}{The output file from the Victor run}
}
\value{
A dataframe of the assay results is returned
}
\description{
Retrieves and cleans assay results from Victor ouput file
}
