% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata}
\usage{
get_metadata(file)
}
\arguments{
\item{file}{The output file from the Victor run}
}
\value{
A list of metadata is returned
}
\description{
Retrieves metadata from Victor ouput file
}
