% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{import_victor_file}
\alias{import_victor_file}
\title{Import Victor File}
\usage{
import_victor_file(filename)
}
\arguments{
\item{filename}{The output file from the Victor run}
}
\value{
Returns a list of assay and metadata from a Victor file
}
\description{
Imports and manipulates the victor file
}
