% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{nfkb_cols}
\alias{nfkb_cols}
\title{NFkB Expression Test columns}
\format{An object of class \code{character} of length 2.}
\usage{
nfkb_cols
}
\description{
Use these columns when testing NFkB expression.
\describe{
  \item{Control}{control well}
  \item{10 ng/ul TNFa}{experiemental well}
}
}
\keyword{datasets}
