% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_file}
\alias{parse_file}
\title{Parse File}
\usage{
parse_file(input_file, experiment_name)
}
\arguments{
\item{input_file}{DESCRIPTION.}

\item{experiment_name}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Parse the input file
}
