% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{select_data}
\alias{select_data}
\title{Select Data}
\usage{
select_data(data, columns, rows)
}
\arguments{
\item{data}{The assay data object.}

\item{rows}{The rows to select.}

\item{colums}{The rows to select.}
}
\value{
Returns a subsetted data object.
}
\description{
Selects columns and rows from assay data to analyze
}
