% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{select_rows}
\alias{select_rows}
\title{Select Rows}
\usage{
select_rows(data, rows)
}
\arguments{
\item{data}{The assay data object.}

\item{rows}{The rows to select.}
}
\value{
Returns a data object with rows selected.
}
\description{
Selects rows from assay data to analyze
}
