% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{sre_cols}
\alias{sre_cols}
\title{SRE Expression Test columns}
\format{An object of class \code{character} of length 4.}
\usage{
sre_cols
}
\description{
SRE Expression Test columns
}
\keyword{datasets}
